fix bug that updates store without changes
Previously the store compared the old values with the new ones. The
issue is that object comparisons would always return false even if their properties
were the same so now the store turns them into strings first. Given that
the data handled by the store is not expected to become too complex in
the future this is a satisfying solution.